Added map taller
